Move seriazliers into separete module
